Game: split up Gamelogic into smaller parts
